+++ docs/src/components/CVVInput/index.js
@@ -1,0 +1,3 @@
+import CVVInput from './CVVInput';
+
+export default CVVInput;