+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
+## 5.4.0 (2022-03-21)
+### Added
+- Field name on onFocus callback to CVVInput
+- Docs: CVVInput example on docs
+- Docs: CardNumberInput example on docs
+- Add name to Autocomplete onFocus
+
 ## 5.3.0 (2022-03-07)
 ### Added
 - Add name and value to onFocus callback in `TextInput` component
 - Expose CVVInput component
 - Expose CardNumberInput
+
 ## 5.1.1 (2020-10-06)
 ### Changed
 - react and react-dom as peerDependencies